Combine Strategy and StrategyConfig
