+++ job/strategies/popularity.py
@@ -1,4 +1,3 @@
-from abc import ABCMeta, abstractmethod
 from dataclasses import dataclass
 from typing import Any, Dict, List
 
@@ -7,27 +6,30 @@ import pandas as pd
 
 from db.mappings.model import ModelType
 from db.mappings.recommendation import Rec
+from job.strategies.templates.strategy import Strategy
 from sites.templates.site import Site
 
 
 @dataclass
-class Strategy(metaclass=ABCMeta):
-    @abstractmethod
+class Popularity(Strategy):
+    """
+    Default popularity-model site configs.
+    """
+
+    # this is a number of days; will only recommend articles within the past X days
+    popularity_window: int
+
     def fetch_data(site: Site, interactions_data: pd.DataFrame) -> List[Dict[str, Any]]:
         pass
 
-    @abstractmethod
     def preprocess_data(site: Site, article_data: List[Dict[str, str]], interactions_data: pd.DataFrame) -> pd.DataFrame:
         pass
 
-    @abstractmethod
     def generate_embeddings(train_data: pd.DataFrame) -> np.ndarray:
         pass
 
-    @abstractmethod
     def generate_recommendations(train_embeddings: np.ndarray, train_data: pd.DataFrame) -> List[Rec]:
         pass
 
-    @abstractmethod
     def save_recommendations(site: Site, recs: List[Rec], model_type: ModelType) -> None:
         pass
